requirements.txt
aiohttp==3.7.3
aiohttp-socks==0.8.4
aiolimiter==1.1.1
aiosqlite==0.20.0
anyio==4.5.2
APScheduler==3.10.4
async-timeout==3.0.1
asyncio==3.4.3
attrs==24.3.0
backports.zoneinfo==0.2.1
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
chardet==3.0.4
cryptography==44.0.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
multidict==6.1.0
nest-asyncio==1.6.0
propcache==0.2.0
pycparser==2.22
pypng==0.20220715.0
python-dotenv==1.0.1
python-socks==2.6.1
python-telegram-bot==21.6
pytz==2024.2
qrcode==7.4.2
six==1.17.0
sniffio==1.3.1
tornado==6.4.2
typing-extensions==4.12.2
tzlocal==5.2
yarl==1.15.2
